+++ changelog.md
@@ -6,6 +6,7 @@
 - Added **levels**, an interface for them and methods for `BaseLevel`. `InitEgridenComponents()` now creates a default level.
 - Added **free layers** which allow you to place and draw objects anywhere on the screen according to their XY coordinates.
 - Added methods for deleting and moving Gobjects within a grid layer: `l.DeleteAt()`, `l.MoveGobjectTo()`.
+- Removed `baseGobjectWithoutScripts`, you can use the new `OnDrawFunc` and `OnUpdateFunc` fields to assign scripts, which are nil by default.
 - Added `l.IsXYwithinBounds()`.
 - Renamed `l.AddObject` to `l.AddGobject`.
 - Updated to Ebitengine v2.7.3.
@@ -16,7 +17,7 @@
 - Removed `Gobject.DoesDrawScriptOverwriteSprite()` and introduced `Gobject.DrawSprite()`. Use the latter inside the `Gobject.OnDraw()` for combine sprite and custom drawing.
     - `Gobject.OnDraw()` returned now function needs a layer pointer to achieve this. This also allows for custom drawing to be interconnected with the layer fields.
 - Removed the non-functional `OnCreate()` as its too much dependent on how you want to implement custom Gobjects and how they are created. The creator/build function should replace this.
-- Updated to go v1.22.2 and Ebitengine v2.7.2
+- Updated to go v1.22.2 and Ebitengine v2.7.3
 
 ## v0.1.0